FSDEX-2247 Blueshift Amm investigation

+++ src/main/java/fi/blueshift/lib/domain/graphqlDto/TokenDto.java
@@ -53,7 +53,7 @@ public class TokenDto {
      * token decimals
      */
     @GraphQLInputField
-    Long decimals;
+    Integer decimals;
 
     /**
      * token total supply
